move compat stuff to own file
