+++ python_modules/dagster/dagster/_config/pythonic_config/__init__.py
@@ -21,13 +21,7 @@ from typing import (
     cast,
 )
 
-import pydantic
 from pydantic.fields import FieldInfo
-
-try:
-    from pydantic_core import PydanticUndefined  # type: ignore
-except:
-    PydanticUndefined = None
 from typing_extensions import Annotated, TypeAlias, TypeGuard, get_args, get_origin
 
 from dagster import (
@@ -83,9 +77,6 @@ except ImportError:
         pass
 
 
-USING_PYDANTIC_2 = int(pydantic.__version__.split(".")[0]) >= 2
-
-
 from abc import ABC, abstractmethod
 
 from pydantic import BaseModel
@@ -105,6 +96,13 @@ from dagster._core.definitions.resource_definition import (
 )
 from dagster._core.storage.io_manager import IOManager, IOManagerDefinition
 
+from .pydantic_compat_layer import (
+    USING_PYDANTIC_2,
+    ModelFieldCompat,
+    PydanticUndefined,
+    model_config,
+    model_fields,
+)
 from .typing_utils import BaseConfigMeta, BaseResourceMeta, LateBoundTypesForResourceTypeChecking
 from .utils import safe_is_subclass
 
@@ -114,93 +112,6 @@ INTERNAL_MARKER = "__internal__"
 
 # ensure that this ends with the internal marker so we can do a single check
 assert CACHED_METHOD_FIELD_SUFFIX.endswith(INTERNAL_MARKER)
-
-
-class ModelFieldCompat:
-    """Wraps a Pydantic model field to provide a consistent interface for accessing
-    metadata and annotations between Pydantic 1 and 2.
-    """
-
-    def __init__(self, field):
-        self.field = field
-
-    @property
-    def annotation(self):
-        return self.field.annotation
-
-    @property
-    def metadata(self):
-        return getattr(self.field, "metadata", None)
-
-    @property
-    def alias(self):
-        return self.field.alias
-
-    @property
-    def default(self):
-        return self.field.default
-
-    @property
-    def description(self):
-        # Pydantic 1.x
-        field_info = getattr(self.field, "field_info", None)
-        if field_info:
-            return field_info.description
-
-        # Pydantic 2.x
-        return getattr(self.field, "description", None)
-
-    def is_required(self):
-        # Pydantic 2.x
-        if hasattr(self.field, "is_required"):
-            return self.field.is_required()
-
-        # Pydantic 1.x
-        return self.field.required
-
-    @property
-    def discriminator(self):
-        # Pydantic 2.x
-        if hasattr(self.field, "discriminator"):
-            return self.field.discriminator
-
-        # Pydantic 1.x
-        return getattr(self.field, "discriminator_key", None)
-
-
-def model_fields(model) -> Dict[str, ModelFieldCompat]:
-    """Returns a dictionary of fields for a given pydantic model, wrapped
-    in a compat class to provide a consistent interface between Pydantic 1 and 2.
-    """
-    fields = getattr(model, "model_fields", None)
-    if not fields:
-        fields = getattr(model, "__fields__")
-
-    return {k: ModelFieldCompat(v) for k, v in fields.items()}
-
-
-class Pydantic1ConfigWrapper:
-    """Config wrapper for Pydantic 1 style model config, which provides a
-    Pydantic 2 style interface for accessing mopdel config values.
-    """
-
-    def __init__(self, config):
-        self._config = config
-
-    def get(self, key):
-        return getattr(self._config, key)
-
-
-def model_config(model: Type[BaseModel]):
-    """Returns the config for a given pydantic model, wrapped such that it has
-    a Pydantic 2-style interface for accessing config values.
-    """
-    # Pydantic 2.x
-    if hasattr(model, "model_config"):
-        return getattr(model, "model_config")
-
-    # Pydantic 1.x
-    return Pydantic1ConfigWrapper(getattr(model, "__config__"))
 
 
 class MakeConfigCacheable(BaseModel):
